Clang Format / Git Check Tests
This patch adds tests to Travis CI for Clang Format and Git
Check (whitespace). Formatting is version dependent, and
thus, Clang Format 4.0 is currently being used. If an issue
is found, the results are printed to the Travis CI log to
provide the person submitting a PR with a list of changes
that need to be made.

In addition to adding Clang Format and Git Check tests,
this patch also adds tests for Clang 3.9 and Clang 4.0 as
well as genrally cleaning up the Travis CI script, getting
it ready for possibly adding Code Coverage tests as well
as Static / Dynamic Analysis tests.

Finally, Travis CI has been updated to trusty to add tests
for Clang 4.0+. Trusty does not support 3.7 so this has
been disabled, and 3.9 Release has been commented out for
now as it segfaults during test (likely an issue with
optimizations that has been there for a while now).

Signed-off-by: Rian Quinn <“rianquinn@gmail.com”>